// src/components/CardHeader/CardHeader.types.ts
export interface CardHeaderProps {
  icon?: React.ReactNode
  heading?: React.ReactNode
  description?: React.ReactNode
  badge?: React.ReactNode
  menu?: React.ReactNode
  showIcon?: boolean
  showTitle?: boolean
  showDescription?: boolean
  showBadge?: boolean
  showMenu?: boolean
  className?: string
}